requirements.txt
streamlit
cryptography
pycryptodome
